#2493 Rationalise date validation in `RangeValidatorFunction`
Extract methods for comparing dates into a utility class.
They have already been used in TG Ports, but it is not clear whether RangeValidatorFunction
should be a public API -- better use an existing utility class.
